Un-parenting from CDH root pom
This jar file has a vastly different release schedule from that of the
rest of CDH, and it's only vaguely related in code. Not having CDH as
a parent also means that I don't have to wait for an official CDH 6
release to move forward with updating third party libraries.