Creation and use of the "RedactionPolicyParseException"
Here we create an exception for this class, and return it in place of
Jackson json-parsing exceptions. This is a Good Thing because it hides
Jackson entirely from the caller. We're also shading / relocating
Jackson, meaning that the full Jackson class names are also super long
and should be private.