Whitelisting fasterxml.jackson for jar relocation
Previously I was attempting to relocate everything, and then
specifically exclude a few jar paths. After my second obscure failure
in this paradigm I'm giving up and instead only specifically including
what needs to be relocated right now, which is just jackson.  In the
future we'll need to be careful to relocate any future-added
dependencies.